A lot done
So, I managed to update the way the lists look for the categories and
perks, so it can look a lot more appealing and professional. Right now,
we need to only create a URL of a XML. Also, I edited the window for the
categories screen. Still need to make it look better and more appealing.